backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.44
asyncpg==0.30.0
alembic==1.16.5
pydantic==2.11.9
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
werkzeug==3.0.1
python-multipart==0.0.20
httpx==0.28.1
groq==0.31.1
tavily-python==0.7.11
crewai==0.177.0
crewai-tools==0.75.0
apscheduler==3.11.0
email-validator==2.1.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
